move function into example code
